Added bullets to three.js
The bullet vertices sent from three js are just two, representing a line
segment rather than a shape.  Since the vertices never change, I just
used a constant definition of vertices on the client side for bullets.
We require at least a shape with three vertices, but we defined a box
with four of size 10x10.